Add files via upload
Quality assessment is the measurement of the quality of
healthcare services. A quality assessment measures the
difference between expected and actual performance to
identify opportunities for improvement. Performance standards
can be established for most dimensions of quality, such as
technical competence, effectiveness, efficiency, safety, and
coverage. Where standards are established, a quality assessment
measures the level of compliance with standards. For
dimensions of quality where standards are more difficult to
identify, such as continuity of care or accessibility, a quality
assessment describes the current level of performance with
the objective of improving it.
A quality assessment frequently combines various data
collection methods to overcome the intrinsic biases of each
method alone. These methods include direct observation of
patient provider encounters, staff interview, patient focus
group, record review, and facility inspection. The assessment
is often the initial step in a larger process which may include
providing feedback to health workers on performance, training
and motivating staff to undertake quality improvements, and
designing solutions to bridge the quality gap.
This case study describes how local staff in Malawi
collected, analyzed and interpreted quality assessment data to
identify opportunities to improve family planning services.